+++ src/index.ts
@@ -1,10 +1,15 @@
 // Anyhow: index.ts
 
+import {Logger} from "./types"
+import parser from "./parser"
 import _ from "lodash"
+import {libSetup} from "./setup"
+
+// Chalk (colorized console output).
 let chalk = null
 
 /**
- * This is the main (and only) class of the Anyhow library.
+ * This is the main class of the Anyhow library.
  * @example const logger = require("anyhow")
  */
 class Anyhow {
@@ -18,10 +23,29 @@ class Anyhow {
     // --------------------------------------------------------------------------
 
     /**
+     * Internal object defining a Logger.
+     */
+    private _logger: Logger = null
+
+    /**
+     * Getter for _lib, to be used by external modules.
+     */
+    get logger(): Logger {
+        return this._logger
+    }
+
+    /**
+     * For compatibility only, now returns the logger's name.
+     */
+    get lib(): string {
+        return this._logger ? this._logger.name : "none"
+    }
+
+    /**
      * Helper to check if [[setup]] was already called and logger is ready to log.
      */
     get isReady(): boolean {
-        if (this._log) {
+        if (this._logger) {
             return true
         }
 
@@ -29,46 +53,10 @@ class Anyhow {
     }
 
     /**
-     * Private logging function. This will be set on [[setup]].
-     */
-    private _log: Function
-
-    /**
-     * Name of the current library being used, populated on setup().
-     * Possible values are console, winston, bunyan and pino.
-     */
-    private _lib: string = null
-
-    /**
-     * Getter for _lib, to be used by external modules.
-     */
-    get lib(): string {
-        return this._lib
-    }
-
-    /**
      * Name of the current running app. This will be used when logging
      * uncaught exceptions, rejections and in similar cases.
      */
     appName: string
-
-    /**
-     * Messages will be compacted (spaces and breaks removed), default is true.
-     * Set to false to log original values including spaces.
-     */
-    compact: boolean = true
-
-    /**
-     * Prepend logged messages with a timestamp on the format YY-MM-DD hh:mm:ss.
-     * default is false.
-     */
-    timestamp: boolean = false
-
-    /**
-     * Log error stack traces? Default is false. Use it with care!
-     * Set to true to add stack traces to the log output.
-     */
-    errorStack: boolean = false
 
     /**
      * Include level (INFO, WARN, ERROR etc...) on the output message when
@@ -81,12 +69,6 @@ class Anyhow {
      * it's [[info]], [[warn]] and [[error]], so [[debug]] won't log anything.
      */
     levels: string[] = ["info", "warn", "error"]
-
-    /**
-     * Set the separator between arguments when generating logging messages.
-     * Default separator is a pipe symbol " | ".
-     */
-    separator: string = " | "
 
     /**
      * Default console logging styles to be used in case the `chalk` module is installed.
@@ -102,14 +84,6 @@ class Anyhow {
         /** Display error messages in bold red. */
         error: ["red", "bold"]
     }
-
-    /**
-     * A function(arrayOfObjects) that should be called to preprocess the arguments.
-     * This is useful if you wish to remove or obfuscate data before generating the
-     * logging message. Can either mutate the passed arguments or return the
-     * processed array as a result.
-     */
-    preprocessor: Function
 
     /**
      * Function to catch and log uncaught exceptions, set by [[uncaughtExceptions]].
@@ -165,6 +139,66 @@ class Anyhow {
         }
     }
 
+    // PARSER PROPERTIES
+    // --------------------------------------------------------------------------
+
+    /**
+     * Messages will be compacted (spaces and breaks removed), default is true.
+     * Set to false to log original values including spaces.
+     */
+    get compact(): boolean {
+        return parser.compact
+    }
+    set compact(value: boolean) {
+        parser.compact = value
+    }
+
+    /**
+     * Prepend logged messages with a timestamp on the format YY-MM-DD hh:mm:ss.
+     * default is false.
+     */
+    get timestamp(): boolean {
+        return parser.timestamp
+    }
+    set timestamp(value: boolean) {
+        parser.timestamp = value
+    }
+
+    /**
+     * Log error stack traces? Default is false. Use it with care!
+     * Set to true to add stack traces to the log output.
+     */
+    get errorStack(): boolean {
+        return parser.errorStack
+    }
+    set errorStack(value: boolean) {
+        parser.errorStack = value
+    }
+
+    /**
+     * Set the separator between arguments when generating logging messages.
+     * Default separator is a pipe symbol " | ".
+     */
+    get separator(): string {
+        return parser.separator
+    }
+    set separator(value: string) {
+        parser.separator = value
+    }
+
+    /**
+     * A function(arrayOfObjects) that should be called to preprocess the arguments.
+     * This is useful if you wish to remove or obfuscate data before generating the
+     * logging message. Can either mutate the passed arguments or return the
+     * processed array as a result.
+     */
+    get preprocessor(): Function {
+        return parser.preprocessor
+    }
+    set preprocessor(value: Function) {
+        parser.preprocessor = value
+    }
+
     // LOGGING METHODS
     // --------------------------------------------------------------------------
 
@@ -177,14 +211,14 @@ class Anyhow {
     log(level: string, args: any | any[]): string {
         if (this.levels.indexOf(level.toLowerCase()) < 0) return null
 
-        let message = _.isString(args) ? args : this.getMessage(args)
+        let message = _.isString(args) ? args : parser.getMessage(args)
 
         // If setup was not called yet, defaults to console logging and emit warning.
         if (!this.isReady) {
             console.warn("Anyhow: please call Anyhow's setup() on your application init. Will default to console.log() for now.")
             this.console(level, message)
         } else {
-            this._log(level, message)
+            this._logger.log(level, message)
         }
 
         return message
@@ -196,7 +230,7 @@ class Anyhow {
     debug = (...args: any[]): string => {
         if (this.levels.indexOf("debug") < 0) return null
         if (args.length < 1) return
-        let message = this.getMessage(args)
+        let message = parser.getMessage(args)
         return this.log("debug", message)
     }
 
@@ -206,7 +240,7 @@ class Anyhow {
     info = (...args: any[]): string => {
         if (this.levels.indexOf("info") < 0) return null
         if (args.length < 1) return
-        let message = this.getMessage(args)
+        let message = parser.getMessage(args)
         return this.log("info", message)
     }
 
@@ -216,7 +250,7 @@ class Anyhow {
     warn = (...args: any[]): string => {
         if (this.levels.indexOf("warn") < 0) return null
         if (args.length < 1) return
-        let message = this.getMessage(args)
+        let message = parser.getMessage(args)
         return this.log("warn", message)
     }
 
@@ -226,7 +260,7 @@ class Anyhow {
     error = (...args: any[]): string => {
         if (this.levels.indexOf("error") < 0) return null
         if (args.length < 1) return
-        let message = this.getMessage(args)
+        let message = parser.getMessage(args)
         return this.log("error", message)
     }
 
@@ -240,7 +274,7 @@ class Anyhow {
     console = (level: string, args: any): string => {
         if (this.levels.indexOf(level.toLowerCase()) < 0) return null
 
-        let message = _.isString(args) ? args : this.getMessage(args)
+        let message = _.isString(args) ? args : parser.getMessage(args)
 
         // Add level to the output?
         if (this.levelOnConsole) {
@@ -253,6 +287,14 @@ class Anyhow {
         // Check if console supports the passed level. Defaults to "log".
         if (console[level] && level != "debug") {
             logMethod = console[level]
+        }
+
+        try {
+            if (chalk === null) {
+                chalk = require("chalk")
+            }
+        } catch (ex) {
+            chalk = false
         }
 
         // Is chalk installed? Use it to colorize the messages.
@@ -278,250 +320,14 @@ class Anyhow {
     }
 
     /**
-     * Setup will try to load compatible loggers, and fall back
-     * to the console if nothing was found. You can also force
-     * a specific library to be loaded passing the object directly or
-     * as string "winston|bunyan|pino|console". Set "none" to disable logging.
-     * @param lib Optional, force a specific library to be used. If not passed, will try winston then bunyan then console.
+     * Setup will try to load compatible loggers, and fall back to the console
+     * if nothing was found. Will try using libraries on this order:
+     * winston, bunyan, pino, gcloud, console.
+     * @param lib Optional, force a specific library or Logger to be used.
      * @param options Additional options to be passed to the underlying logging library.
      */
-    setup = (lib?: string | any, options?: any): void => {
-        let found = false
-        let winston, bunyan, pino
-
-        // Set defaults.
-        lib = lib || "console"
-        options = options || {}
-
-        // Passed "none"? This will effectively disable logging.
-        if (lib == "none") {
-            this._lib = null
-            this._log = function () {
-                return false
-            }
-
-            return
-        }
-
-        // Passed logger directly? Figure out which one it is based on object info.
-        if (_.isObject(lib)) {
-            if (lib.constructor.name == "DerivedLogger" && lib.format && lib.level) {
-                winston = lib
-                lib = "winston"
-            } else if (lib.constructor.name == "Logger" && lib.fields) {
-                bunyan = lib
-                lib = "bunyan"
-            } else if (lib.constructor.name == "Pino" && lib.levels) {
-                bunyan = lib
-                lib = "pino"
-            } else {
-                console.warn("Anyhow.setup", "Passed object was not recognized as Winston / Bunyan / Pino.")
-            }
-        }
-
-        // First try Winston. It will check if a Winston logger was passed directly
-        // as `lib`, or use default logger if passed as string.
-        if (lib == "winston") {
-            try {
-                if (!winston) {
-                    winston = require("winston")
-                }
-
-                this._lib = "winston"
-                this._log = function (level, message) {
-                    winston.log({level: level, message: message})
-                }
-
-                found = true
-            } catch (ex) {
-                /* istanbul ignore next */
-                console.error("Anyhow.setup", "Could not load winston", ex)
-            }
-        }
-
-        // Then Bunyan. It will check if a Bunyan logger was passed directly
-        // as `lib`, or create a default logger if passed as string.
-        if (lib == "bunyan") {
-            try {
-                if (!options.name) {
-                    options.name = "Anyhow"
-                }
-
-                if (!bunyan) {
-                    bunyan = require("bunyan").createLogger(options)
-                }
-
-                this._lib = "bunyan"
-                this._log = function (level, message) {
-                    bunyan[level](message)
-                }
-
-                found = true
-            } catch (ex) {
-                /* istanbul ignore next */
-                console.error("Anyhow.setup", "Could not load bunyan", ex)
-            }
-        }
-
-        // Then Pino. It will check if a Pino logger was passed directly
-        // as `lib`, or create a default logger if passed as string.
-        if (lib == "pino") {
-            try {
-                if (!options.name) {
-                    options.name = "Anyhow"
-                }
-
-                if (!pino) {
-                    pino = require("pino")()
-                }
-
-                this._lib = "pino"
-                this._log = function (level, message) {
-                    pino[level](message)
-                }
-
-                found = true
-            } catch (ex) {
-                /* istanbul ignore next */
-                console.error("Anyhow.setup", "Could not load pino", ex)
-            }
-        }
-
-        // No logging libraries found? Fall back to console.
-        if (!found) {
-            try {
-                if (!chalk) {
-                    chalk = require("chalk")
-                }
-            } catch (ex) {}
-
-            this._lib = "console"
-            this._log = (level, message) => this.console(level, message)
-        }
-    }
-
-    /**
-     * Gets a nice, readable message out of the passed arguments, which can be of any type.
-     * @param args Any single or collection of objects that will be transformed to a message string.
-     * @returns Human readable string taken out of the parsed arguments.
-     */
-    getMessage = (...args: any | any[]): string => {
-        if (!_.isArray(args)) {
-            args = [args]
-        }
-
-        args = _.flattenDeep(args)
-
-        // Add timestamp to the output?
-        if (this.timestamp) {
-            const padLeft = (v) => {
-                return v < 10 ? "0" + v.toString() : v.toString()
-            }
-
-            // Get date elements.
-            const now = new Date()
-            let year: any = now.getUTCFullYear().toString()
-            let month: any = now.getUTCMonth() + 1
-            let day: any = now.getUTCDate()
-            let hour: any = now.getUTCHours()
-            let minute: any = now.getUTCMinutes()
-            let second: any = now.getUTCSeconds()
-
-            // Append timestamp to the message.
-            let timestamp = `${padLeft(year.substring(2))}-${padLeft(month)}-${padLeft(day)} ${padLeft(hour)}:${padLeft(minute)}:${padLeft(second)}`
-            args.unshift(timestamp)
-        }
-
-        // If the preprocessor returns a value, use it as the new args.
-        if (this.preprocessor) {
-            let processedArgs = this.preprocessor(args)
-            args = processedArgs ? processedArgs : args
-        }
-
-        // Return single string log message.
-        return this.parseArgsForMessage(args).join(this.separator)
-    }
-
-    /**
-     * Used by [[getMessage]] to parse and return the individual log strings
-     * out of the passed arguments. Might run recursively.
-     * @param args Array of arguments to be parsed.
-     */
-    private parseArgsForMessage(args: any[]): string[] {
-        let result = []
-
-        // Parse all arguments and stringify objects. Please note that fields defined
-        // on the `removeFields` setting won't be added to the message.
-        for (let arg of args) {
-            try {
-                if (arg != null) {
-                    let stringified = ""
-
-                    try {
-                        if (_.isArray(arg)) {
-                            result = result.concat(this.parseArgsForMessage(arg))
-                        } else if (_.isError(arg)) {
-                            const arrError = []
-
-                            // Add error information separately.
-                            if (arg.friendlyMessage) {
-                                arrError.push(arg.friendlyMessage)
-                            }
-                            if (arg.reason) {
-                                arrError.push(arg.reason)
-                            }
-                            if (arg.code) {
-                                arrError.push(arg.code)
-                            }
-
-                            /* istanbul ignore else */
-                            if (arg.message) {
-                                arrError.push(arg.message)
-                            }
-
-                            // Errors from axios.
-                            if (arg.response && arg.response.data) {
-                                if (arg.response.data.message) {
-                                    arrError.push(arg.response.data.message.toString())
-                                }
-                                if (arg.response.data.error) {
-                                    arrError.push(arg.response.data.error.toString())
-                                }
-                            }
-
-                            // Only add stack traces if `errorStack` is set.
-                            if (arg.stack && this.errorStack) {
-                                arrError.push(arg.stack)
-                            }
-
-                            stringified = arrError.join(this.separator)
-                        } else if (_.isObject(arg)) {
-                            stringified = JSON.stringify(arg, null, 2)
-                        } else {
-                            stringified = arg.toString()
-                        }
-                    } catch (ex) {
-                        /* istanbul ignore next */
-                        stringified = arg.toString()
-                    }
-
-                    // Check if a valid message was taken, and if it needs to be compacted.
-                    /* istanbul ignore else */
-                    if (typeof stringified != "undefined" && stringified !== null) {
-                        if (this.compact) {
-                            stringified = stringified.replace(/(\r\n|\n|\r)/gm, "").replace(/  +/g, " ")
-                        }
-
-                        result.push(stringified)
-                    }
-                }
-            } catch (ex) {
-                /* istanbul ignore next */
-                console.error("Anyhow.parseArgsForMessage", ex)
-            }
-        }
-
-        return result
+    setup = (lib?: "winston" | "bunyan" | "pino" | "gcloud" | "console" | "none" | Logger, options?: any): void => {
+        libSetup(this, lib, options)
     }
 }